get careful about equality and hashing
I'm seeing building programs take a big enough
chunk of time, and of that the deduplication of
instructions, that this seems worth trying now.

Wanted to upload this to look at it, but
no speed difference.  The real fix is probably
to add something like

   void programKey(std::vector<uint32_t>*) const;

to write out a key instead of running the
whole shader->program() and using the hash
of its instructions.  Kind of dodgy to use
a hash as an ID anyway.

+++ src/core/SkVM.h
@@ -319,7 +319,7 @@ namespace skvm {
         M(select) M(bytes) M(pack)            \
     // End of SKVM_OPS
 
-    enum class Op : uint8_t {
+    enum class Op : int {
     #define M(op) op,
         SKVM_OPS(M)
     #undef M
